fix(editors): normalize notification service import casing
Fixes #44
